+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Features
 
 1. [#209](https://github.com/InfluxCommunity/influxdb3-java/pull/209) Add query function returning row as map.
-1. [#238](https://github.com/InfluxCommunity/influxdb3-java/pull/238): Support fast writes without waiting for WAL
+2. [#238](https://github.com/InfluxCommunity/influxdb3-java/pull/238): Support fast writes without waiting for WAL
    persistence:
    - New write option (`WriteOptions.noSync`) added: `true` value means faster write but without the confirmation that
      the data was persisted. Default value: `false`.
@@ -13,6 +13,7 @@
    - Long precision string values added from v3 HTTP API: `"nanosecond"`, `"microsecond"`, `"millisecond"`,
      `"second"` (
      in addition to the existing `"ns"`, `"us"`, `"ms"`, `"s"`).
+3. [#241](https://github.com/InfluxCommunity/influxdb3-java/pull/241): Some default options will be used from a getter.
 
 ### Bug Fixes